feat(docs): migrate theme reference
