+++ docs/components/BorderRadius.js
@@ -17,15 +17,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { css } from '@emotion/core';
-import { ThemeProvider } from 'emotion-theming';
-import Text from '../../../../src/components/Text';
-import { theme as themes } from '../../../../src';
+import Text from '../../src/components/Text';
+import { theme as themes } from '../../src';
 
 const Box = styled('div')`
   ${({ theme, size }) => css`
-    width: ${theme.iconSizes[size]};
-    height: ${theme.iconSizes[size]};
-    border-radius: ${theme.borderRadius.kilo};
+    width: ${theme.spacings.tera};
+    height: ${theme.spacings.tera};
+    border-radius: ${theme.borderRadius[size]};
     background-color: ${theme.colors.r300};
     margin-right: ${theme.spacings.mega};
   `};
@@ -39,39 +38,37 @@ const Wrapper = styled('div')`
   `};
 `;
 
-const IconSizeSize = styled('span')`
+const BorderRadiusSize = styled('span')`
   ${({ theme }) => css`
     color: ${theme.colors.n500};
   `};
 `;
 
-const IconSizeName = styled(Text)`
+const BorderRadiusName = styled(Text)`
   ${({ theme }) => css`
     margin-left: ${theme.spacings.kilo};
     color: ${theme.colors.n500};
   `};
 `;
 
-const IconSize = ({ size }) => (
-  <ThemeProvider theme={themes.circuit}>
-    <Wrapper>
-      <Box size={size} />
-      <div>
-        <Text as="span">{size}</Text>
-        <IconSizeSize>
-          <IconSizeName size={Text.KILO} as="span">
-            {themes.circuit.iconSizes[size]}
-          </IconSizeName>
-        </IconSizeSize>
-      </div>
-    </Wrapper>
-  </ThemeProvider>
+const BorderRadius = ({ size }) => (
+  <Wrapper>
+    <Box size={size} />
+    <div>
+      <Text as="span">{size}</Text>
+      <BorderRadiusSize>
+        <BorderRadiusName size={Text.KILO} as="span">
+          {themes.circuit.borderRadius[size]}
+        </BorderRadiusName>
+      </BorderRadiusSize>
+    </div>
+  </Wrapper>
 );
 
-IconSize.propTypes = {
+BorderRadius.propTypes = {
   // eslint-disable-next-line
   theme: PropTypes.object.isRequired,
   size: PropTypes.string.isRequired
 };
 
-export default IconSize;
+export default BorderRadius;
