+++ docs/components/index.js
@@ -28,5 +28,8 @@ export { default as Swatch } from './Swatch';
 export { default as Type } from './Type';
 export { default as Spacing } from './Spacing';
 export { default as MediaQueriesTable } from './MediaQueriesTable';
+export { default as BorderWidth } from './BorderWidth';
+export { default as BorderRadius } from './BorderRadius';
+export { default as IconSize } from './IconSize';
 export { Grid, Row, Col } from './Grid';
 export { default as MDXStyles } from './MDXStyles';
